Extract huc8 geojsons reliably
To generate all the huc8 geojsons (around 2266) we need to ensure
a couple of things:

1. All the nhdflowline features are present in the nhdplusv2 table.
2. All the wbdhu8 boundaries are present in the nhdplushhr table.

The sql queries had to be modified because the original sql queries were
generating a ton of warnings. These log warnings used up a lot of space on the
database docker container and raised the "No space left on the device"
error.

The huc8 geojson code generation was modified slightly to only run for
geojsons that are not already present in the output dir. Even though
docker stats lists that docker is using all the cpus assigned to docker it
takes quite a bit of time to generate these geojsons. For instance, it
took almost an hour to generate 75 geojsons on my laptop (actually
docker with 6 cpus and 12 gb memory). Coincidentally, the battery went
from 100% charged to 48% charged in 1.5 hours!

To recreate all the geojsons just delete all the geojsons with rm
*.json. However, if you are wary of doing that, just do mv huc8-extracts
huc8-extracts-old and the script will recreate the folder. By the way,
this is a good idea...especially if you wish to compare if the geojsons
generated are different in any possible way.

Closes #48

fix: remove commented out code blocks.

As per Lewis's recommendation, removed commented out code blocks. The
geojson generation finished but learned that there are some duplicate
huc8 boundaries (181 of them)! You can find them in the wbdhu8 table
with `select huc8,array_agg(distinct "name") "name", array_agg(distinct
tnmid) tnmid, count(*) cnt from wbdhu8 group by huc8 having count(*)
> 1`.

fixed quoting inconsistencies