Update notebooks with .persist
This keeps the data in memory once it is loaded. This does not change
the runtime behavior of NetCDF much, but has a big impact on Zarr and
Kerchunk examples.

Also add a Kerchunk Combined example. Its behavior is notably different.
This is likely because we're using xr.open_dataset rather than xr.open_mfdataset.